Improving sample data (#92)
* Requires recent pooch

The syntax used depends on recent release of pooch. It's always a good
idea to define a range of versions on dependencies.

* Initiating sample_data with wc_6beam

* Cleaning imports not used

* Diverting estimateion 6-bem notebook to sample_data()

* Use explicit lidarwind package name

It's not a good idea to abreviate packages that are not a community
standard. It's OK for xarray, numpy, etc, but not for lidarwid. There is
a big chance that another package in a completely different area using
the same abreviation.

* Adding wc_long_dbs group of sample_data

* style

I was running a new repo without pre-commit, so I missed some minor
adjustments.

* fix: typo

* Using utilities.sample_data() instead

* Using sample_data in turbulence_6beam_data.ipynb

* Using sample_data() on reading_long_dbs_rendered

* Cleaning import not used

* dbs_scans_rendered notebook with sample_data()

* merging_6beam_rendered notebook using sample_data()

* fix: Typo when moving from lst

We should not use lst instead of explicit lidarwind package name. It's
OK to do that with largely used packages such as xarray, numpy, and
others, but not lst.

* Caching samples within GA

* Updating GA's checkout